Update task + automation folder structure
