+++ booru/automation/tag/__init__.py
