+++ booru/automation/tag/metadata.py
