+++ booru/automation/tag/tag_automation.py
