+++ booru/automation/tag/tags.py
