+++ booru/models/automation.py
@@ -189,7 +189,7 @@ class NSFWAutomationRecord(models.Model):
 
 # Hook into the Post save method to re-add the post to be scanned
 from django.db.models.signals import post_save
-from booru.tasks import perform_automation
+from booru.tasks.tag_automation import perform_automation
 
 def post_save_post(sender, instance, created, **kwargs):
     """Removes the post from the automation records."""
